move dialog.dart to route.dart
