+++ lib/web_ui/lib/src/engine.dart
@@ -145,7 +145,6 @@ export 'engine/scene_painting.dart';
 export 'engine/scene_view.dart';
 export 'engine/semantics/accessibility.dart';
 export 'engine/semantics/checkable.dart';
-export 'engine/semantics/dialog.dart';
 export 'engine/semantics/focusable.dart';
 export 'engine/semantics/heading.dart';
 export 'engine/semantics/image.dart';
@@ -154,6 +153,7 @@ export 'engine/semantics/label_and_value.dart';
 export 'engine/semantics/link.dart';
 export 'engine/semantics/live_region.dart';
 export 'engine/semantics/platform_view.dart';
+export 'engine/semantics/route.dart';
 export 'engine/semantics/scrollable.dart';
 export 'engine/semantics/semantics.dart';
 export 'engine/semantics/semantics_helper.dart';
