+++ lib/web_ui/lib/src/engine/semantics/semantics.dart
@@ -20,7 +20,6 @@ import '../vector_math.dart';
 import '../window.dart';
 import 'accessibility.dart';
 import 'checkable.dart';
-import 'dialog.dart';
 import 'focusable.dart';
 import 'heading.dart';
 import 'image.dart';
@@ -29,6 +28,7 @@ import 'label_and_value.dart';
 import 'link.dart';
 import 'live_region.dart';
 import 'platform_view.dart';
+import 'route.dart';
 import 'scrollable.dart';
 import 'semantics_helper.dart';
 import 'tappable.dart';
